Refactor for easier testing
